+++ src-api/org/seasar/mayaa/engine/processor/ProcessorTreeWalker.java
@@ -17,14 +17,12 @@ package org.seasar.mayaa.engine.processor;
 
 import java.util.Map;
 
-import org.seasar.mayaa.engine.specification.serialize.NodeReferenceResolverFinder;
-
 /**
  * プロセッサツリーを操作する。
  * @author Masataka Kurihara (Gluegent, Inc.)
  */
 public interface ProcessorTreeWalker
-        extends NodeReferenceResolverFinder {
+    {
 
     /**
      * プロセッサ実行スコープに、登録スクリプト変数を提供する。