+++ src-impl/org/seasar/mayaa/impl/engine/specification/NodeTreeWalkerImpl.java
@@ -22,8 +22,6 @@ import java.util.List;
 import java.util.Objects;
 
 import org.seasar.mayaa.engine.specification.NodeTreeWalker;
-import org.seasar.mayaa.engine.specification.serialize.NodeReferenceResolver;
-import org.seasar.mayaa.impl.engine.specification.serialize.NodeSerializeController;
 
 /**
  * @author Taro Kato (Gluegent, Inc.)
@@ -148,10 +146,6 @@ public class NodeTreeWalkerImpl implements NodeTreeWalker {
 
     public boolean isOnTemplate() {
         return _onTemplate;
-    }
-
-    public NodeReferenceResolver findNodeResolver() {
-        return NodeSerializeController.currentInstance();
     }
 
     /* (non-Javadoc)